server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
text_generation==0.3.0
openai==0.27.8
tiktoken
python-multipart
pinecone-client
langchain==0.0.345